Add files via upload
Once the code is written and run then a request for scraping is sent to your website . If the request is approved, the server will send the desired data, allowing you to read the HTML  page. The code then automatically analyses the HTML or XML page, finds and parses the desired data.